# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Coordina ADG API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitation codes the caller has generated or can manage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Generate a role-based invitation code
      description: Generates a shareable invitation code scoped to a role (and optional province/center). The email is unknown at creation time; the recipient provides it when registering. Coordinators generate codes for department heads and prospectors; department heads for teachers.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreated"
        "403":
          $ref: "#/components/responses/Error"
  /invitations/{id}:
    delete:
      operationId: revokeInvitation
      tags: [invitations]
      summary: Revoke a pending invitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Revoked
  /invitations/{id}/resend:
    post:
      operationId: resendInvitation
      tags: [invitations]
      summary: Renew an invitation code (extend its expiry)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreated"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    StatusQuery:
      name: status
      in: query
      schema:
        type: string
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required: [message]
...
    Role:
      type: string
      enum:
        - superadmin
        - coordinator
        - prospector
        - department_head
        - teacher
...
    Invitation:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        email:
          type: string
          nullable: true
          description: Email of the user who redeemed the code; null until redeemed.
        role:
          $ref: "#/components/schemas/Role"
        provinceId:
          type: integer
          nullable: true
        centerId:
          type: integer
          nullable: true
        status:
          type: string
        expiresAt:
          type: string
          format: date-time
        usedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, code, role, status, expiresAt]
    CreateInvitationInput:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
        provinceId:
          type: integer
          nullable: true
        centerId:
          type: integer
          nullable: true
        expiresInHours:
          type: integer
          default: 72
      required: [role]
    InvitationCreated:
      type: object
      properties:
        invitation:
          $ref: "#/components/schemas/Invitation"
        inviteUrl:
          type: string
      required: [invitation, inviteUrl]
...